adds new audiobook test and reorganizes tests
